requirements.txt
# Avocado-CloudTest functional requirements
# SSH lib (avocado.utils.remote)
fabric>=1.11.0
# multiplexer (avocado.core.tree)
PyYAML>=3.11
# .tar.xz support (avocado.utils.archive)
# REST client (avocado.core.restclient)
requests>=1.2.3
# stevedore for loading "new style" plugins
stevedore>=1.8.0; python_version >= '2.7'
# stevedore-1.11.0 won't support python2.6 anymore
stevedore>=1.8.0,<=1.10.0; python_version < '2.7'
# Additional python 2.6 specific requirements for avocado and unittests (backports)
argparse>=1.3.0; python_version < '2.7'
logutils>=0.3.3; python_version < '2.7'
importlib>=1.0.3; python_version < '2.7'
unittest2>=1.0.0
aexpect>=1.2.0
pystache>=0.5.4
bottle>=0.12
ipaddr>=2.1
cuisine>=0.7
numpy>=1.13
pandas>=0.20
scipy>=0.19
networkx>=1.11
os-client-config>=1.28.0
python-cinderclient>=3.1.0
python-glanceclient>=2.8.0
python-keystoneclient>=3.13.0
python-neutronclient>=6.5.0
python-novaclient>=9.1.0
python-openstackclient>=3.12.0
aodhclient>=0.9.0
python-ceilometerclient>=2.9.0
paramiko>=2.3.1
